# docs-v2/30-api/openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Teranga API",
    "description": "African Event Management Platform — REST API for events, registrations, badges, and users",
    "version": "0.1.0",
    "contact": {
      "name": "Teranga Team",
      "email": "dev@teranga.events"
    }
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "Firebase JWT",
        "description": "Firebase ID token obtained via Firebase Auth SDK"
      },
      "ApiKeyAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "terk_<env>_<40chars>_<4chksum>",
        "description": "Organization API key (T2.3). Issued via /admin/organizations/[id]?tab=api-keys."
      }
    },
...
  },
  "paths": {
...
    "/v1/registrations/": {
      "post": {
        "summary": "Register for an event",
        "tags": [
          "Registrations"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/v1/registrations/me": {
      "get": {
        "summary": "Get my registrations",
        "tags": [
          "Registrations"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/v1/registrations/me/event/{eventId}": {
      "get": {
        "summary": "Get my current registration for an event (any non-terminal status)",
        "tags": [
          "Registrations"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "eventId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/v1/registrations/event/{eventId}": {
      "get": {
        "summary": "List registrations for an event",
        "tags": [
          "Registrations"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "eventId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/v1/registrations/{registrationId}": {
      "delete": {
        "summary": "Cancel a registration",
        "tags": [
          "Registrations"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "registrationId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      },
      "patch": {
        "summary": "Update registration status (promote or cancel)",
        "tags": [
          "Registrations"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "registrationId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/v1/registrations/{registrationId}/cancel": {
      "post": {
        "summary": "Cancel a registration (POST alias for DELETE)",
        "tags": [
          "Registrations"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "registrationId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/v1/registrations/{registrationId}/cancel-pending": {
      "post": {
        "summary": "Cancel my own pending_payment registration (release the slot)",
        "tags": [
          "Registrations"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "registrationId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/v1/registrations/{registrationId}/approve": {
      "post": {
        "summary": "Approve a pending registration",
        "tags": [
          "Registrations"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "registrationId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/v1/registrations/{registrationId}/waitlist-position": {
      "get": {
        "summary": "Get a participant's position on the waitlist (FIFO within ticket type)",
        "tags": [
          "Registrations"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "registrationId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/v1/registrations/checkin": {
      "post": {
        "summary": "Check-in via QR code",
        "tags": [
          "Registrations"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/v1/registrations/bulk-cancel": {
      "post": {
        "summary": "Bulk-cancel many registrations (per-row failures collected)",
        "tags": [
          "Registrations"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/v1/registrations/bulk-approve": {
      "post": {
        "summary": "Bulk-approve many waitlisted registrations",
        "tags": [
          "Registrations"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local dev"
    }
  ],
...
}